requirements.txt
appdirs==1.4.3
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
attrs==19.1.0
bcrypt==3.1.6
black==19.3b0
blinker==1.4
Brlapi==0.6.7
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
command-not-found==0.3
cryptography==2.6.1
cupshelpers==1.0
dbus-python==1.2.12
defer==1.0.6
defusedxml==0.5.0
diff-match-patch==20121119
distro==1.3.0
distro-info===0.21ubuntu4
Django==3.0.1
django-allauth==0.41.0
django-bootstrap-datepicker-plus==3.0.5
django-cors-headers==3.0.2
django-crispy-forms==1.8.1
django-extensions==2.0.7
django-grappelli==2.13.1
django-import-export==1.0.1
django-model-utils==3.1.2
django-rest-knox==4.1.0
django-widget-tweaks==1.4.2
djangorestframework==3.11.0
duplicity==0.8.4
entrypoints==0.3
et-xmlfile==1.0.1
fasteners==0.12.0
future==0.16.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-oauth==1.0.1
gunicorn==19.9.0
httplib2==0.11.3
idna==2.8
jdcal==1.4
kazam==1.4.5
keyring==18.0.1
keyrings.alt==3.1.1
language-selector==0.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.10.0
macaroonbakery==1.2.3
Mako==1.0.7
MarkupSafe==1.1.0
monotonic==1.5
mysqlclient==1.3.10
netifaces==0.10.4
oauth==1.0.1
oauthlib==2.1.0
odfpy==1.3.6
olefile==0.46
openpyxl==2.5.6
paramiko==2.6.0
pbr==5.4.4
pexpect==4.6.0
Pillow==6.2.1
protobuf==3.6.1
psycopg2-binary==2.8.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.34.0
PyJWT==1.6.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
python-apt==1.9.0+ubuntu1
python-dateutil==2.7.3
python-debian==0.1.36
python-decouple==3.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.2
pyxdg==0.26
PyYAML==5.1.2
reportlab==3.5.23
requests==2.22.0
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
rsa==4.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.12.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
tablib==0.12.1
testresources==2.0.1
toml==0.10.0
ubuntu-advantage-tools==19.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.25.3
usb-creator==0.3.7
vboxapi==1.0
virtualenv==15.1.0
wadllib==1.3.3
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.3.2
